Split person test files
